Move to new design that handles cluster failure correctly
We simplify the design. Before we were constrained in an unatural way by
having a 'Protocol' module that defined the operations for a single
server. This seems the natural, or ideal way to compose the operations
from a single memcache server into a cluster of memcache servers.

Sadly, it didn't work out that well. Duplicating operations like `get`
at the server and cluster level and making error handling and failover
difficult.

Now, we treat servers as pipes we can send requests to and receive
responses from. Operations like `get` only exist at the cluster level,
removing code and simplifying error handling.

The exception hierachy is now unified and much simpler too.